Charts

LightweightCharts.LWCChart — Type

LWCChart

The base type that contains the necessary information for visualizing a single chart (and its plugins).

See also: lwc_show, lwc_save.

Line

LightweightCharts.LWCCharts.lwc_line — Function

lwc_line([, timestamps], values::Vector{Real}; kw...) -> LWCChart

Creates a LWCChart object that contains line chart information. The timestamps can be passed as Vector{Integer} of Unix time or Vector{TimeType}.

Wrapper function for Line.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

line_color::String	Random color	Line color.

line_style::LWC_LINE_STYLES	LWC_SOLID (LWC_DOTTED, LWC_DASHED, LWC_LARGE_DASHED, LWC_SPARSE_DOTTED)	Line style.

line_width::Int64	1	Line width.

line_type::LWC_LINE_TYPES	LWC_SIMPLE (LWC_STEP, LWC_CURVED)	Line type.

line_visible::Bool	true	Line visibility.

point_markers_visible::Bool	false	Displaying markers on line nodes.

point_markers_radius::Float64	4.0	Size of markers.

crosshair_marker_visible::Bool	true	Cursor crosshair visibility.

crosshair_marker_radius::Float64	4.0	Size of crosshair.

crosshair_marker_border_color::String	"#758696"	Border color of crosshair.

crosshair_marker_background_color::String	"#4c525e"	Background color of crosshair.

crosshair_marker_border_width::Float64	2.0	Border width of the crosshair.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

Example

using LightweightCharts

t_values = round(Int64, time()) .+ collect(1:500)
x_values = map(x -> sin(x / 10), 1:500)

chart = lwc_line(
    t_values,
    x_values;
    label_name = "lwc_line",
    line_color = "#02c39a",
    line_width = 4,
    line_style = LWC_SOLID,
    line_type = LWC_CURVED,
    price_scale_id = LWC_RIGHT,
)

lwc_save("line_example.html", chart)

Baseline

LightweightCharts.LWCCharts.lwc_baseline — Function

lwc_baseline([, timestamps], values::Vector{Real}; kw...) -> LWCChart

Creates a LWCChart object that contains baseline chart information. The timestamps can be passed as Vector{Integer} of Unix time or Vector{TimeType}.

Wrapper function for Baseline.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

base_value::LWCBaseValue	LWCBaseValue("price", 0.0)	The value relative to which larger or smaller values will be colored in the appropriate color.

top_fill_color1::String	"rgba(38, 166, 154, 0.28)"	The first color of the upper area.

top_fill_color2::String	"rgba(38, 166, 154, 0.05)"	The second color of the upper area.

top_line_color::String	"rgba(38, 166, 154, 1)"	Color of the upper area line.

bottom_fill_color1::String	"rgba(239, 83, 80, 0.05)"	The first color of the lower area.

bottom_fill_color2::String	"rgba(239, 83, 80, 0.28)"	The second color of the lower area.

bottom_line_color::String	"rgba(239, 83, 80, 1)"	Color of the lower area line.

line_width::Int64	1	Line width.

line_style::LWC_LINE_STYLES	LWC_SOLID (LWC_DOTTED, LWC_DASHED, LWC_LARGE_DASHED, LWC_SPARSE_DOTTED)	Line style.

line_type::LWC_LINE_TYPES	LWC_SIMPLE (LWC_STEP, LWC_CURVED)	Line type.

line_visible::Bool	true	Line visibility.

point_markers_visible::Bool	false	Displaying markers on line nodes.

point_markers_radius::Float64	4.0	Size of markers.

crosshair_marker_visible::Bool	true	Cursor crosshair visibility.

crosshair_marker_radius::Float64	4.0	Size of crosshair.

crosshair_marker_border_color::String	"#758696"	Border color of crosshair.

crosshair_marker_background_color::String	"#4c525e"	Background color of crosshair.

crosshair_marker_border_width::Float64	2.0	Border width of the crosshair.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

Example

using LightweightCharts

t_values = round(Int64, time()) .+ collect(1:500)
x_values = map(x -> sin(x / 10) + cos(x / 5), 1:500)

chart = lwc_baseline(
    t_values,
    x_values;
    label_name = "lwc_baseline",
    base_value = LWCBaseValue("price", 0.0),
    line_style = LWC_SOLID,
    line_type = LWC_CURVED,
    line_width = 3,
    precision = 4,
    price_scale_id = LWC_RIGHT,
)

lwc_save("baseline_example.html", chart)

Area

LightweightCharts.LWCCharts.lwc_area — Function

lwc_area([, timestamps], values::Vector{Real}; kw...) -> LWCChart

Creates a LWCChart object that contains area chart information. The timestamps can be passed as Vector{Integer} of Unix time or Vector{TimeType}.

Wrapper function for Area.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

top_color::String	"rgba(46, 220, 135, 0.4)"	Color of the upper part of the area.

bottom_color::String	"rgba(40, 221, 100, 0)"	Color of the lower part of the area.

invert_filled_area::Bool	false	Inverted display of colors.

line_color::String	Random color	Line color.

line_style::LWC_LINE_STYLES	LWC_SOLID (LWC_DOTTED, LWC_DASHED, LWC_LARGE_DASHED, LWC_SPARSE_DOTTED)	Line style.

line_width::Int64	1	Line width.

line_type::LWC_LINE_TYPES	LWC_SIMPLE (LWC_SIMPLE, LWC_STEP, LWC_CURVED)	Line type.

point_markers_visible::Bool	false	Displaying markers on line nodes.

point_markers_radius::Float64	4.0	Size of markers.

crosshair_marker_visible::Bool	true	Cursor crosshair visibility.

crosshair_marker_radius::Float64	4.0	Size of crosshair.

crosshair_marker_border_color::String	"#758696"	Border color of crosshair.

crosshair_marker_background_color::String	"#4c525e"	Background color of crosshair.

crosshair_marker_border_width::Float64	2.0	Border width of the crosshair.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

Example

using Dates
using LightweightCharts

t_values = round(Int64, time()) .+ collect(1:500)
x_values = map(x -> x % 20, 1:500)

chart = lwc_area(
    t_values,
    x_values;
    label_name = "lwc_area",
    line_color = "#00b4d8",
    top_color = "#90e0ef",
    bottom_color = "rgba(133, 242, 240, 0)",
    line_style = LWC_SOLID,
    line_type = LWC_STEP,
    line_width = 2,
    precision = 3,
    price_scale_id = LWC_RIGHT,
)

lwc_save("area_example.html", chart)

Histogram

LightweightCharts.LWCCharts.lwc_histogram — Function

lwc_histogram([, timestamps], values::Vector{Real}; kw...) -> LWCChart

Creates a LWCChart object that contains histogram chart information. The timestamps can be passed as Vector{Integer} of Unix time or Vector{TimeType}.

Wrapper function for Histogram.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

color::String	Random color	Histogram color.

base::Real	0.0	The value relative to which the larger or smaller histogram values will be located.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

Example

using LightweightCharts

t_values = round(Int64, time()) .+ collect(1:500)
x_values = map(x -> x * cos(x / 10), 1:500)

chart = lwc_histogram(
    t_values,
    x_values;
    label_name = "lwc_histogram",
    base = 0.0,
    color = "rgba(240, 113, 103, 0.5)",
    price_scale_id = LWC_RIGHT,
)

lwc_save("histogram_example.html", chart)

OHLC Candles

LightweightCharts.LWCChartData.LWCCandle — Type

LWCCandle(time::Real, open::Real, high::Real, low::Real, close::Real)
LWCCandle(time::TimeType, open::Real, high::Real, low::Real, close::Real)

Representation of candlestick data for lwc_candlestick and lwc_bar methods.

Fields

time::Int64

open::Float64

high::Float64

low::Float64

close::Float64

Candlestick

LightweightCharts.LWCCharts.lwc_candlestick — Function

lwc_candlestick(data::Vector{LWCCandle}; kw...) -> LWCChart

Creates a LWCChart object that contains candlesticks chart information.

Wrapper function for Candlestick.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

up_color::String	"#26a69a"	Color of bullish candle (increasing).

down_color::String	"#ef5350"	Color of bearish candle (decreasing).

wick_visible::Bool	true	Wick visibility.

border_visible::Bool	true	Candle borders visibility.

border_color::String	"#378658"	Candle border color.

border_up_color::String	"#26a69a"	Boder color of bullish candle.

border_down_color::String	"#ef5350"	Boder color of bearish candle.

wick_color::String	"#737375"	Wick color.

wick_up_color::String	"#26a69a"	Wick color of bullish candle.

wick_down_color::String	"#ef5350"	Wick color of bearish candle.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

lwc_candlestick(arg...; kw...) -> LWCChart

Takes as input individual vectors with the corresponding candlestick values.

Arguments

timestamps::Vector{Union{Real,TimeType}}

open::Vector{Real}

high::Vector{Real}

low::Vector{Real}

close::Vector{Real}

lwc_candlestick(data::Vector{Tuple{D,O,H,L,C}}; kw...) -> LWCChart

Takes a single vector with tuples containing the corresponding candlestick values. Here D is a Real or TimeType and O,H,L,C are Reals.

Example

using Dates
using LightweightCharts

open_time = now() .+ Second.(1:500)
x_values = map(x -> sin(2rand() + x / 10), 1:500)

chart = lwc_candlestick(
    LWCCandle.(
        open_time,
        x_values,
        x_values .+ rand(500),
        x_values .- rand(500),
        [x_values[2:end]..., x_values[begin]],
    );
    label_name = "lwc_candlestick",
    up_color = "#52a49a",
    down_color = "#de5e57",
    border_visible = false,
    price_scale_id = LWC_RIGHT,
)

lwc_save("candlestick_example.html", chart)

Bar

LightweightCharts.LWCCharts.lwc_bar — Function

wc_bar(data::Vector{LWCCandle}; kw...) -> LWCChart

Creates a LWCChart object that contains candlesticks chart information.

Wrapper function for Bar.

Keyword arguments

Name::Type	Default (Posible values)	Description

price_scale_id::LWC_PRICE_SCALE_ID	LWC_LEFT (LWC_RIGHT)	Y-axis display side.

label_name::String	""	Chart name.

visible::Bool	true	Chart visibility.

precision::Int64	2	Number of decimal places for y-axis values.

up_color::String	"#26a69a"	Color of bullish candle (increasing).

down_color::String	"#ef5350"	Color of bearish candle (decreasing).

open_visible::Bool	true	Open tick visibility.

thin_bars::Bool	true	Thin bars.

plugins::Vector{LWCPlugin}	LWCPlugin[]	Additional plugins.

lwc_bar(arg...; kw...) -> LWCChart

Takes as input individual vectors with the corresponding candlestick values.

Arguments

timestamps::Vector{Union{Real,TimeType}}

open::Vector{Real}

high::Vector{Real}

low::Vector{Real}

close::Vector{Real}

wc_bar(data::Vector{Tuple{D,O,H,L,C}}; kw...) -> LWCChart

Takes a single vector with tuples containing the corresponding candlestick values. Here D is a Real or TimeType and O,H,L,C are Reals.

Example

using Dates
using LightweightCharts

open_time = now() .+ Second.(1:500)
x_values = map(x -> sin(2rand() + x / 10), 1:500)

chart = lwc_bar(
    LWCCandle.(
        open_time,
        x_values,
        x_values .+ rand(500),
        x_values .- rand(500),
        [x_values[2:end]..., x_values[begin]],
    );
    label_name = "lwc_bar",
    up_color = "#52a49a",
    down_color = "#de5e57",
    price_scale_id = LWC_RIGHT,
)

lwc_save("bar_example.html", chart)